feat: Added option to create a package group
When creating a new web component, there is a new option to select an
existing package group or to create a new one.

Each package group have a unique custom tag prefix and organization.

Also was updated the default test 'hello-world' package, it was
recreated on packages/foundations/atoms/hello-world.

The .env file is now included in git.